Major refactor which uses Email::Sender instead of Email::Stuff
Email::Stuff is built on Email::Send which is deprecated in favor of
Email::Sender. This is a major refactor which basically re-implements
the entire module on top of Email::Sender. This is a breaking change. I
tried my best to keep the interface the same (email { ... }), but the
configuration has changed drastically.