Give access to MIME fields to the user for the attachments
This patch doesn't break retrocompatibility. When passing an hashref
to the attachment, pass the hash to MIME::Entity, complaining if Path
is missing.

If a string is passed, the original routine is used.

The example in the documentation has been changed to show a possible
use (embedding CID images in HTML mails).